remove subgraph.handlebars.yaml and other devenv improvements
