+++ src/mappings/Avatar/datasource.yaml
@@ -1,0 +1,7 @@
+entities:
+  - AvatarContract
+eventHandlers:
+  - event: SendEther(uint256, address)
+    handler: handleSendEth
+  - event: ReceiveEther(address, uint256)
+    handler: handleReceiveEth
