+++ src/mappings/ContributionReward/datasource.yaml
@@ -1,0 +1,20 @@
+entities:
+  - ContributionRewardRedeemReputation
+  - ContributionRewardRedeemNativeToken
+  - ContributionRewardRedeemExternalToken
+  - ContributionRewardRedeemEther
+  - ContributionRewardProposalExecuted
+  - ContributionRewardNewContributionProposal
+eventHandlers:
+  - event: RedeemReputation(address, bytes32, address, int256)
+    handler: handleRedeemReputation
+  - event: RedeemNativeToken(address, bytes32, address, uint256)
+    handler: handleRedeemNativeToken
+  - event: RedeemExternalToken(address, bytes32, address, uint256)
+    handler: handleRedeemExternalToken
+  - event: RedeemEther(address, bytes32 , address ,uint256)
+    handler: handleRedeemEther
+  - event: ProposalExecuted(address, bytes32, int256)
+    handler: handleProposalExecuted
+  - event: NewContributionProposal(address, bytes32, address, bytes32, int256, uint256[5], address, address)
+    handler: handleNewContributionProposal
