+++ src/mappings/GenesisProtocol/mapping.ts
