+++ src/mappings/NativeToken/datasource.yaml
@@ -1,0 +1,22 @@
+abis:
+  - DAOToken
+entities:
+  - TokenHolder
+  - TokenMint
+  - TokenMintFinished
+  - TokenTransfer
+  - TokenApproval
+  - TokenBurn
+eventHandlers:
+  - event: Mint(address, uint256)
+    handler: handleMint
+  - event: Burn(address, uint256)
+    handler: handleBurn
+  - event: Approval(address, uint256)
+    handler: handleApproval
+  - event: MintFinished()
+    handler: handleMintFinished
+  - event: Transfer(address, address, uint256)
+    handler: handleTransfer
+  - event: OwnershipTransferred(address, address)
+    handler: handleOwnershipTransferred
