+++ src/mappings/UController/datasource.yaml
@@ -1,0 +1,25 @@
+abis:
+  - UController
+  - Avatar
+  - Reputation
+  - DAOToken
+entities:
+  - UControllerOrganization
+  - UControllerScheme
+  - UControllerGlobalConstraint
+  - UControllerRegisterScheme
+  - UControllerUnregisterScheme
+  - UControllerUpgradeController
+  - UControllerAddGlobalConstraint
+  - UControllerRemoveGlobalConstraint
+eventHandlers:
+  - event: RegisterScheme(address, address, address)
+    handler: handleRegisterScheme
+  - event: UnregisterScheme(address, address, address)
+    handler: handleUnregisterScheme
+  - event: UpgradeController(address, address, address)
+    handler: handleUpgradeController
+  - event: AddGlobalConstraint(address, bytes32, uint8, address)
+    handler: handleAddGlobalConstraint
+  - event: RemoveGlobalConstraint(address, uint256, bool, address)
+    handler: handleRemoveGlobalConstraint
