--- src/mappings/contributionReward/ContributionReward.graphql
@@ -1,81 +1,0 @@
-type ContributionRewardRedeemReputation @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    proposalId: Bytes!
-    amount: BigInt!
-}
-
-type ContributionRewardRedeemNativeToken @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    proposalId: Bytes!
-    amount: BigInt!
-}
-
-type ContributionRewardRedeemExternalToken @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    proposalId: Bytes!
-    amount: BigInt!
-}
-
-type ContributionRewardRedeemEther @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    proposalId: Bytes!
-    amount: BigInt!
-}
-
-type ContributionRewardProposalResolved @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    proposalId: Bytes!
-    passed: Boolean
-}
-
-type ContributionRewardNewContributionProposal @entity {
-    txHash: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    descriptionHash: Bytes!
-    externalToken: Bytes!
-    votingMachine: Bytes!
-    proposalId: Bytes!
-    reputationReward: BigInt!
-    nativeTokenReward: BigInt!
-    ethReward: BigInt!
-    externalTokenReward: BigInt!
-    periods: BigInt!
-    periodLength: BigInt!
-}
-
-type ContributionRewardProposal @entity {
-    proposalId: ID!
-    contract: Bytes!
-    avatar: Bytes!
-    beneficiary: Bytes!
-    descriptionHash: Bytes!
-    externalToken: Bytes!
-    votingMachine: Bytes!
-    reputationReward: BigInt!
-    nativeTokenReward: BigInt!
-    ethReward: BigInt!
-    externalTokenReward: BigInt!
-    periods: BigInt!
-    periodLength: BigInt!
-    executedAt: BigInt
-    alreadyRedeemedReputationPeriods: BigInt
-    alreadyRedeemedNativeTokenPeriods: BigInt
-    alreadyRedeemedEthPeriods: BigInt
-    alreadyRedeemedExternalTokenPeriods: BigInt
-}
