--- src/mappings/daotoken/DAOToken.graphql
@@ -1,47 +1,0 @@
-type TokenContract @entity {
-    id: ID!
-    address: Bytes!
-    totalSupply: BigInt!
-    owner: Bytes!
-}
-
-type TokenHolder @entity {
-    contract: Bytes!
-    address: Bytes!
-    balance: BigInt!
-}
-
-type TokenMint @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    to: Bytes!
-    amount: BigInt!
-}
-
-type TokenBurn @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    burner: Bytes!
-    amount: BigInt!
-}
-
-type TokenMintFinished @entity {
-    txHash: Bytes!
-    contract: Bytes!
-}
-
-type TokenTransfer @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    from: Bytes!
-    to: Bytes!
-    value: BigInt!
-}
-
-type TokenApproval @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    owner: Bytes!
-    spender: Bytes!
-    value: BigInt!
-}
