--- src/mappings/genesisProtocol/GenesisProtocol.graphql
@@ -1,66 +1,0 @@
-type GenesisProtocolProposal @entity {
-  proposalId: ID!
-  submittedTime: BigInt!
-  proposer: Bytes!
-  daoAvatarAddress: Bytes!
-  numOfChoices: BigInt
-  executionState: Int
-  state: Int
-  decision: BigInt
-  executionTime: BigInt
-  totalReputation: BigInt
-}
-
-type GenesisProtocolVote @entity {
-  avatarAddress: Bytes!
-  voterAddress: Bytes!
-  reputation: BigInt!
-  voteOption: BigInt!
-  proposalId: GenesisProtocolProposal!
-}
-
-type GenesisProtocolStake @entity {
-  avatarAddress: Bytes!
-  stakerAddress: Bytes!
-  prediction: BigInt!
-  stakeAmount: BigInt!
-  proposalId: GenesisProtocolProposal!
-}
-
-type GenesisProtocolRedemption @entity {
-  rewardId: GenesisProtocolReward!
-  proposalId: ID!
-  redeemer: Bytes!
-}
-
-type GenesisProtocolReward @entity {
-  id: ID!
-  type: GenesisProtocolRewardType
-  amount: BigInt!
-}
-
-type GenesisProtocolExecuteProposal @entity {
-  txHash: ID!
-  contract: Bytes!
-  proposalId: Bytes!
-  organization:Bytes!
-  decision: BigInt!
-  totalReputation: BigInt!
-}
-
-type GenesisProtocolGPExecuteProposal @entity {
-  txHash: ID!
-  contract: Bytes!
-  proposalId: Bytes!
-  executionState: Int
-}
-
-enum GenesisProtocolRewardType @entity {
-  beneficiaryEth
-  beneficiaryNativeToken
-  beneficiaryReputation
-  beneficiaryExternalToken
-  gpRep
-  gpGen
-  gpBounty
-}
