--- src/mappings/reputation/Reputation.graphql
@@ -1,25 +1,0 @@
-type ReputationContract @entity {
-    id: ID!
-    address: Bytes!
-    totalSupply: BigInt!
-}
-
-type ReputationHolder @entity {
-    contract: Bytes!
-    address: Bytes!
-    balance: BigInt!
-}
-
-type ReputationMint @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    address: Bytes!
-    amount: BigInt!
-}
-
-type ReputationBurn @entity {
-    txHash: Bytes!
-    contract: Bytes!
-    address: Bytes!
-    amount: BigInt!
-}
