--- subgraph.handlebars.yaml
@@ -1,198 +1,0 @@
-# template file for generating subgraph.yaml based on configuration
-
-specVersion: 0.0.1
-schema:
-  file: ./schema.graphql
-dataSources:
-- kind: ethereum/contract
-  name: UController
-  source:
-    address: "{{addresses.UController}}"
-    abi: UController
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/uController/UController.ts
-    entities:
-    - UControllerOrganization
-    - UControllerScheme
-    - UControllerGlobalConstraint
-    - UControllerRegisterScheme
-    - UControllerUnregisterScheme
-    - UControllerUpgradeController
-    - UControllerAddGlobalConstraint
-    - UControllerRemoveGlobalConstraint
-    abis:
-    - name: UController
-      file: ./abis/UController.json
-    - name: Avatar
-      file: ./abis/Avatar.json
-    - name: Reputation
-      file: ./abis/Reputation.json
-    - name: DAOToken
-      file: ./abis/DAOToken.json
-    eventHandlers:
-    - event: RegisterScheme(address, address, address)
-      handler: handleRegisterScheme
-    - event: UnregisterScheme(address, address, address)
-      handler: handleUnregisterScheme
-    - event: UpgradeController(address, address, address)
-      handler: handleUpgradeController
-    - event: AddGlobalConstraint(address, bytes32, uint8, address)
-      handler: handleAddGlobalConstraint
-    - event: RemoveGlobalConstraint(address, uint256, bool, address)
-      handler: handleRemoveGlobalConstraint
-
-- kind: ethereum/contract
-  name: Reputation
-  source:
-    address: "{{addresses.Reputation}}"
-    abi: Reputation
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/reputation/Reputation.ts
-    entities:
-    - ReputationContract
-    - ReputationHolder
-    - ReputationMint
-    - ReputationBurn
-    abis:
-    - name: Reputation
-      file: ./abis/Reputation.json
-    eventHandlers:
-    - event: Mint(address, uint256)
-      handler: handleMint
-    - event: Burn(address, uint256)
-      handler: handleBurn
-
-- kind: ethereum/contract
-  name: ContributionReward
-  source:
-    address: "{{addresses.ContributionReward}}"
-    abi: ContributionReward
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/contributionReward/ContributionReward.ts
-    entities:
-    - ContributionRewardRedeemReputation
-    - ContributionRewardRedeemNativeToken
-    - ContributionRewardRedeemExternalToken
-    - ContributionRewardRedeemEther
-    - ContributionRewardProposalExecuted
-    - ContributionRewardNewContributionProposal
-    abis:
-    - name: ContributionReward
-      file: ./abis/ContributionReward.json
-    eventHandlers:
-    - event: RedeemReputation(address, bytes32, address, int256)
-      handler: handleRedeemReputation
-    - event: RedeemNativeToken(address, bytes32, address, uint256)
-      handler: handleRedeemNativeToken
-    - event: RedeemExternalToken(address, bytes32, address, uint256)
-      handler: handleRedeemExternalToken
-    - event: RedeemEther(address, bytes32 , address ,uint256)
-      handler: handleRedeemEther
-    - event: ProposalExecuted(address, bytes32, int256)
-      handler: handleProposalExecuted
-    - event: NewContributionProposal(address, bytes32, address, bytes32, int256, uint256[5], address, address)
-      handler: handleNewContributionProposal
-
-- kind: ethereum/contract
-  name: DAOToken
-  source:
-    address: "{{addresses.DAOToken}}"
-    abi: DAOToken
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/daotoken/DAOToken.ts
-    entities:
-    - TokenHolder
-    - TokenMint
-    - TokenMintFinished
-    - TokenTransfer
-    - TokenApproval
-    - TokenBurn
-    abis:
-    - name: DAOToken
-      file: ./abis/DAOToken.json
-    eventHandlers:
-    - event: Mint(address, uint256)
-      handler: handleMint
-    - event: Burn(address, uint256)
-      handler: handleBurn
-    - event: Approval(address, uint256)
-      handler: handleApproval
-    - event: MintFinished()
-      handler: handleMintFinished
-    - event: Transfer(address, address, uint256)
-      handler: handleTransfer
-    - event: OwnershipTransferred(address, address)
-      handler: handleOwnershipTransferred
-
-- kind: ethereum/contract
-  name: GenesisProtocol
-  source:
-    address: "{{addresses.GenesisProtocol}}"
-    abi: GenesisProtocol
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/genesisProtocol/GenesisProtocol.ts
-    entities:
-      - Proposal
-      - Vote
-      - GPStake
-      - Redemption
-      - Reward
-      - RewardType
-      - ExecuteProposal
-      - GPExecuteProposal
-    abis:
-      - name: GenesisProtocol
-        file: ./abis/GenesisProtocol.json
-    eventHandlers:
-      - event: NewProposal(bytes32, address, uint256, address, bytes32)
-        handler: handleNewProposal
-      - event: Redeem(bytes32, address, address, uint256)
-        handler: handleRedeem
-      - event: RedeemDaoBounty(bytes32, address, address, uint256)
-        handler: handleRedeemDaoBounty
-      - event: RedeemReputation(bytes32, address, address, uint256)
-        handler: handleRedeemReputation
-      - event: VoteProposal(bytes32, address, address, uint256, uint256)
-        handler: handleVoteProposal
-      - event: Stake(bytes32, address, address, uint256, uint256)
-        handler: handleStake
-      - event: GPExecuteProposal(bytes32, uint8)
-        handler: handleGPExecuteProposal
-      - event: ExecuteProposal(bytes32, address, uint256, uint256)
-        handler: handleExecuteProposal
-
-- kind: ethereum/contract
-  name: Avatar
-  source:
-    address: "{{addresses.Avatar}}"
-    abi: Avatar
-  mapping:
-    kind: ethereum/events
-    apiVersion: 0.0.1
-    language: wasm/assemblyscript
-    file: ./src/mappings/avatar/Avatar.ts
-    entities:
-      - AvatarContract
-    abis:
-      - name: Avatar
-        file: ./abis/Avatar.json
-    eventHandlers:
-      - event: SendEther(uint256, address)
-        handler: handleSendEth
-      - event: ReceiveEther(address, uint256)
-        handler: handleReceiveEth
