+++ test/integration/Avatar.spec.ts
@@ -41,8 +41,8 @@ describe('Avatar', () => {
       id: addresses.Avatar.toLowerCase(),
       address: addresses.Avatar.toLowerCase(),
       name: 'TESTDAO',
-      nativeToken: addresses.DAOToken.toLowerCase(),
-      nativeReputation: addresses.Reputation.toLowerCase(),
+      nativeToken: addresses.NativeToken.toLowerCase(),
+      nativeReputation: addresses.NativeReputation.toLowerCase(),
       balance: '1',
       owner: accounts[0].address.toLowerCase(),
     });
