+++ test/integration/Reputation.spec.ts
@@ -11,8 +11,11 @@ describe('Reputation', () => {
     web3 = await getWeb3();
     addresses = getContractAddresses();
     const opts = await getOptions(web3);
-    reputation = new web3.eth.Contract(Reputation.abi, addresses.Reputation, opts);
-
+    reputation = new web3.eth.Contract(
+      Reputation.abi,
+      addresses.NativeReputation,
+      opts,
+    );
   });
 
   async function checkTotalSupply(value) {
