+++ test/integration/UController.spec.ts
@@ -1,4 +1,12 @@
-import { getContractAddresses, getOptions, getWeb3, hashLength, nullParamsHash, padZeros, sendQuery } from './util';
+import {
+  getContractAddresses,
+  getOptions,
+  getWeb3,
+  hashLength,
+  nullParamsHash,
+  padZeros,
+  sendQuery,
+} from './util';
 
 const Avatar = require('@daostack/arc/build/contracts/Avatar.json');
 const DAOToken = require('@daostack/arc/build/contracts/DAOToken.json');
@@ -7,7 +15,6 @@ const TokenCapGC = require('@daostack/arc/build/contracts/TokenCapGC.json');
 const UController = require('@daostack/arc/build/contracts/UController.json');
 
 describe('UController', () => {
-
   let web3;
   let addresses;
   let opts;
@@ -18,46 +25,104 @@ describe('UController', () => {
     web3 = await getWeb3();
     addresses = getContractAddresses();
     opts = await getOptions(web3);
-    uController = new web3.eth.Contract(UController.abi, addresses.UController, opts);
-    reputation = new web3.eth.Contract(Reputation.abi, addresses.Reputation, opts);
-    daoToken = new web3.eth.Contract(DAOToken.abi, addresses.DAOToken, opts);
-
+    uController = new web3.eth.Contract(
+      UController.abi,
+      addresses.UController,
+      opts,
+    );
+    reputation = new web3.eth.Contract(
+      Reputation.abi,
+      addresses.NativeReputation,
+      opts,
+    );
+    daoToken = new web3.eth.Contract(DAOToken.abi, addresses.NativeToken, opts);
   });
 
   it('Sanity', async () => {
     const accounts = web3.eth.accounts.wallet;
     const avatar = await new web3.eth.Contract(Avatar.abi, undefined, opts)
-      .deploy({ data: Avatar.bytecode, arguments: ['Test', daoToken.options.address, reputation.options.address] })
+      .deploy({
+        data: Avatar.bytecode,
+        arguments: [
+          'Test',
+          daoToken.options.address,
+          reputation.options.address,
+        ],
+      })
       .send();
 
-    const tokenCap1 = await new web3.eth.Contract(TokenCapGC.abi, undefined, opts)
+    const tokenCap1 = await new web3.eth.Contract(
+      TokenCapGC.abi,
+      undefined,
+      opts,
+    )
       .deploy({ data: TokenCapGC.bytecode, arguments: [] })
       .send();
 
-    const tokenCap2 = await new web3.eth.Contract(TokenCapGC.abi, undefined, opts)
+    const tokenCap2 = await new web3.eth.Contract(
+      TokenCapGC.abi,
+      undefined,
+      opts,
+    )
       .deploy({ data: TokenCapGC.bytecode, arguments: [] })
       .send();
 
     await avatar.methods.transferOwnership(uController.options.address).send();
     let txs = [];
-    txs.push(await uController.methods.newOrganization(avatar.options.address).send());
-    txs.push(await uController.methods.registerScheme(accounts[2].address,
-                                                      '0x' + padZeros('123', hashLength),
-                                                      '0x' + padZeros('7', 8),
-                                                      avatar.options.address).send());
-    txs.push(await uController.methods.addGlobalConstraint(tokenCap1.options.address,
-                                                           '0x' + padZeros('987', hashLength),
-                                                           avatar.options.address).send());
-    txs.push(await uController.methods.registerScheme(accounts[3].address,
-                                                      '0x' + padZeros('321', hashLength),
-                                                      '0x' + padZeros('8', 8),
-                                                      avatar.options.address).send());
-    txs.push(await uController.methods.addGlobalConstraint(tokenCap2.options.address,
-                                                          '0x' + padZeros('789', hashLength),
-                                                           avatar.options.address).send());
-    txs.push(await uController.methods.unregisterScheme(accounts[3].address, avatar.options.address).send());
-    txs.push(await uController.methods.removeGlobalConstraint(tokenCap2.options.address,
-                                                              avatar.options.address).send());
+    txs.push(
+      await uController.methods.newOrganization(avatar.options.address).send(),
+    );
+    txs.push(
+      await uController.methods
+        .registerScheme(
+          accounts[2].address,
+          '0x' + padZeros('123', hashLength),
+          '0x' + padZeros('7', 8),
+          avatar.options.address,
+        )
+        .send(),
+    );
+    txs.push(
+      await uController.methods
+        .addGlobalConstraint(
+          tokenCap1.options.address,
+          '0x' + padZeros('987', hashLength),
+          avatar.options.address,
+        )
+        .send(),
+    );
+    txs.push(
+      await uController.methods
+        .registerScheme(
+          accounts[3].address,
+          '0x' + padZeros('321', hashLength),
+          '0x' + padZeros('8', 8),
+          avatar.options.address,
+        )
+        .send(),
+    );
+    txs.push(
+      await uController.methods
+        .addGlobalConstraint(
+          tokenCap2.options.address,
+          '0x' + padZeros('789', hashLength),
+          avatar.options.address,
+        )
+        .send(),
+    );
+    txs.push(
+      await uController.methods
+        .unregisterScheme(accounts[3].address, avatar.options.address)
+        .send(),
+    );
+    txs.push(
+      await uController.methods
+        .removeGlobalConstraint(
+          tokenCap2.options.address,
+          avatar.options.address,
+        )
+        .send(),
+    );
     txs = txs.map(({ transactionHash }) => transactionHash);
 
     const { ucontrollerRegisterSchemes } = await sendQuery(`{
@@ -128,8 +193,8 @@ describe('UController', () => {
     expect(ucontrollerOrganizations.length).toEqual(1);
     expect(ucontrollerOrganizations).toContainEqual({
       avatarAddress: avatar.options.address.toLowerCase(),
-      nativeToken: {address: daoToken.options.address.toLowerCase()},
-      nativeReputation: {address: reputation.options.address.toLowerCase()},
+      nativeToken: { address: daoToken.options.address.toLowerCase() },
+      nativeReputation: { address: reputation.options.address.toLowerCase() },
       controller: uController.options.address.toLowerCase(),
     });
 
@@ -230,9 +295,11 @@ describe('UController', () => {
       type: 'Post',
     });
 
-    txs.push((await uController.methods.upgradeController(accounts[4].address,
-                                                          avatar.options.address).send())
-                                                          .transactionHash);
+    txs.push(
+      (await uController.methods
+        .upgradeController(accounts[4].address, avatar.options.address)
+        .send()).transactionHash,
+    );
 
     const { ucontrollerUpgradeControllers } = await sendQuery(`{
       ucontrollerUpgradeControllers {
@@ -251,7 +318,9 @@ describe('UController', () => {
       newController: accounts[4].address.toLowerCase(),
     });
 
-    const { ucontrollerOrganizations: ucontrollerOrganizations2 } = await sendQuery(`{
+    const {
+      ucontrollerOrganizations: ucontrollerOrganizations2,
+    } = await sendQuery(`{
       ucontrollerOrganizations {
         avatarAddress
         nativeToken {
@@ -267,10 +336,9 @@ describe('UController', () => {
     expect(ucontrollerOrganizations2.length).toEqual(1);
     expect(ucontrollerOrganizations2).toContainEqual({
       avatarAddress: avatar.options.address.toLowerCase(),
-      nativeToken: {address: daoToken.options.address.toLowerCase()},
-      nativeReputation: {address: reputation.options.address.toLowerCase()},
+      nativeToken: { address: daoToken.options.address.toLowerCase() },
+      nativeReputation: { address: reputation.options.address.toLowerCase() },
       controller: accounts[4].address.toLowerCase(),
     });
-
   }, 20000);
 });
